BE #7: Initial commit

+++ app/routers/user.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.database import AsyncSessionLocal
@@ -53,11 +53,23 @@ async def get_user(user_id: int, db: AsyncSession = Depends(get_db)):
 async def update_user(
     user_id: int,
     data: UserUpdateRequest,
+    current_user = Depends(get_auth_user),
     db: AsyncSession = Depends(get_db),
 ):
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=403,
+            detail="You can edit only your own profile"
+        )
+
     return await UserService.update_user(db, user_id, data)
 
 
 @router.delete("/{user_id}")
-async def delete_user(user_id: int, db: AsyncSession = Depends(get_db)):
+async def delete_user(user_id: int, current_user = Depends(get_auth_user), db: AsyncSession = Depends(get_db)):
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=403,
+            detail="You can delete only your own profile"
+        )
     return await UserService.delete_user(db, user_id)
